+++ examples/metrics-basic/src/main.rs
@@ -112,7 +112,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
 
     // Create a Gauge Instrument.
     // Note that the Gauge instrument is experimental, and can be changed/removed in the future releases.
-    #[cfg(otel_unstable)]
+    #[cfg(feature = "otel_unstable")]
     {
         let gauge = meter
             .f64_gauge("my_gauge")